pyproject.toml
[project]
name = "nto-ai-25-26-final-baseline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "catboost>=1.2.10",
    "click>=8.3.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.3.3",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.3",
    "pytest>=8.4.2",
    "tqdm>=4.67.3",
]
